refactor: split up the index.js into multiple files
Split up to remove the api calls and put them into their own file
